+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ46ARN2ZNGK3QVEQ1SD40X6"
+        "01KJ4V2WRGCNZV1KG1TSWMV0BX"
     )
     r"""The unique identifier of the data source"""