+++ packages/orq-rc/RELEASES.md
@@ -5949,3 +5949,13 @@ Based on:
 - [python v4.4.0-rc.27] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.27] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.27 - packages/orq-rc
+
+## 2026-02-23 08:53:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.724.0 (2.836.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.28] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.28] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.28 - packages/orq-rc